refactor cli display by creating Response::display
remove bin/cli/display.rs since it is ported
add json field to CommandManager
add global json option to CLI arguments
put certificate loading functions in certificate module
