+++ bin/Cargo.toml
@@ -37,7 +37,6 @@ mio = { version = "^0.8.8", features = ["os-poll", "net"] }
 nix = { version = "^0.27.1", features = ["signal", "fs"] }
 nom = "^7.1.3"
 paw = "^1.0.0"
-prettytable-rs = { version = "^0.10.0", default-features = false }
 serde = { version = "^1.0.188", features = ["derive"] }
 serde_json = "^1.0.107"
 time = "^0.3.29"
@@ -45,7 +44,6 @@ regex = "^1.10.0"
 smol = "^1.3.0"
 tempfile = "^3.8.0"
 termion = "^2.0.1"
-x509-parser = "^0.15.1"
 
 sozu-command-lib = { path = "../command", version = "^0.15.17" }
 sozu-lib = { path = "../lib", version = "^0.15.17" }
