+++ command/Cargo.toml
@@ -42,6 +42,7 @@ serde = { version = "^1.0.188", features = ["derive"] }
 serde_json = "^1.0.107"
 sha2 = "^0.10.8"
 trailer = "^0.1.2"
+prettytable-rs = { version = "^0.10.0", default-features = false }
 pool = "^0.1.4"
 poule = "^0.3.2"
 thiserror = "^1.0.49"
